+++ cadicom/src/main/java/es/uca/cadicom/entity/Tarifa.java
@@ -8,9 +8,9 @@ public class Tarifa {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    private Integer id;
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     private String nombre;
     public String getNombre() { return nombre; }
