+++ cadicom/src/main/java/es/uca/cadicom/entity/Telefono.java
@@ -9,9 +9,9 @@ public class Telefono {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    private Integer id;
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     private String numero;
     public String getNumero() { return numero; }
